+++ src/controllers/productAllergen.controller.js
@@ -26,7 +26,11 @@ export default class ProductAllergenController extends BaseController {
         try {
             const {id} = req.params;
             const result = await this.query("SELECT " + ProductAllergen.visibleFields().join(', ') + " FROM " + ProductAllergen.table() + " WHERE id = ?", id);
-            return response.success('Product Allergen Retrieved successfully', result);
+            const searchProductAllergen = result && result.length === 1 ? new ProductAllergen(result[0]) : undefined;
+            if (!searchProductAllergen) {
+                return response.notFound('Not found product allergen with id ' + id);
+            }
+            return response.success('Product Allergen Retrieved successfully', searchProductAllergen);
         } catch (error) {
             return response.error(error);
         }
@@ -43,7 +47,7 @@ export default class ProductAllergenController extends BaseController {
     
             const result = await this.query("INSERT INTO " + ProductAllergen.table() + " SET ?", productAllergen);
             const id = result.insertId;
-            const data = await this.query("INSERT " + ProductAllergen.visibleFields().join(', ') + " FROM " + ProductAllergen.table() + "WHER id = ?", id);
+            const data = await this.query("SELECT " + ProductAllergen.visibleFields().join(', ') + " FROM " + ProductAllergen.table() + " WHERE id = ?", id);
             return response.success('Product Allergen Created successfully', data);
         } catch (error) {
             return response.error(error);
@@ -54,6 +58,11 @@ export default class ProductAllergenController extends BaseController {
         const response = new HTMLResponse(req, res);
         try {
             const {id} = req.params;
+            const searchResult = await this.query("SELECT " + ProductAllergen.visibleFields().join(', ') + " FROM " + ProductAllergen.table() + " WHERE id = ?", [id]);
+            const searchProduct = searchResult && searchResult.length === 1 ? new ProductAllergen(searchResult[0]) : undefined;
+            if (!searchProduct) {
+                return response.notFound('Not found product allergen with id ' + id);
+            }
             const result = await this.query("DELETE FROM " + ProductAllergen.table() + " WHERE id = ?", id);
             return response.success('Product Allergen Deleted successfully');
         } catch (error) {
